+++ packages/jest-editor-support/index.d.ts
@@ -13,6 +13,7 @@ export interface SpawnOptions {
 }
 
 export interface Options {
+  coverage?: boolean;
   createProcess?(
     workspace: ProjectWorkspace,
     args: string[],