+++ packages/jest-editor-support/src/types.js
@@ -20,6 +20,7 @@ import type {ChildProcess} from 'child_process';
 import type ProjectWorkspace from './project_workspace';
 
 export type Options = {
+  coverage?: boolean,
   createProcess?: (
     workspace: ProjectWorkspace,
     args: Array<string>,