+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Features
 
+* `[jest-editor-support]` Add `coverage` option to runner
+  ([#5836](https://github.com/facebook/jest/pull/5836))
 * `[expect]` Improve output format for mismatchedArgs in mock/spy calls.
   ([#5846](https://github.com/facebook/jest/pull/5846))
 * `[jest-cli]` Add support for using `--coverage` in combination with watch